feat(vouchers): enhance voucher filtering with status and date range support

+++ backend/src/main/java/com/dev/backend/modules/voucher/service/VoucherService.java
@@ -12,5 +12,5 @@ public interface VoucherService {
 
     //VoucherResponse add(Voucher voucher, VoucherRequest request, Integer shopId);
 
-    PageResponse<VoucherResponse> filterVouchers(VoucherFilterRequest request,Integer shopId);
+    PageResponse<VoucherResponse> filterVouchers(VoucherFilterRequest request,Long shopId);
 }
